+++ src/controllers/subCategoryController.js
@@ -1,20 +1,33 @@
-const { APP_URL } = process.env
-const qs = require('querystring')
+const pagination = require('../helpers/pagination')
 const response = require('../helpers/response')
-const { createSubCategoryModel, getAllSubCategoryModel, updateSubCategoryModel, deleteSubCategoryModel, getCountAllSubCategoryModel, getDetailSubCategoryIDModel, getDetailSubCategoryModel, getCountSubCategoryModel } = require('../models/subCategoryModel')
+const {
+  createSubCategoryModel,
+  getAllSubCategoryModel,
+  updateSubCategoryModel,
+  deleteSubCategoryModel,
+  getCountAllSubCategoryModel,
+  getDetailSubCategoryIDModel,
+  getDetailSubCategoryModel,
+  getCountSubCategoryModel
+} = require('../models/subCategoryModel')
 
 module.exports = {
   createSubCategory: async (req, res) => {
     try {
-      const { name, categoryID } = req.body
-      const result = await createSubCategoryModel([name, categoryID])
-      if (result.affectedRows) {
-        return response(res, 'Sub Category has been created', 200, true, { data: req.body })
+      const roleId = req.user.role_id
+      if (roleId === 2) {
+        const { name, categoryID } = req.body
+        const result = await createSubCategoryModel([name, categoryID])
+        if (result.affectedRows) {
+          return response(res, 'Sub Category has been created', 200, true, { data: req.body })
+        } else {
+          return response(res, 'Failed to create Sub Category', 400, false)
+        }
       } else {
-        return response(res, 'Failed to create Sub Category', 400, false)
+        return response(res, 'Forbidden access', 401, false)
       }
     } catch (err) {
-      return response(res, 'Internal server error', 500, false)
+      return response(res, 'Internal server error', 500, false, { error: err.message })
     }
   },
 
@@ -35,29 +48,11 @@ module.exports = {
       }
       const offset = (page - 1) * limit
       const result = await getAllSubCategoryModel([limit, offset])
-      const pageInfo = {
-        count: 0,
-        pages: 0,
-        currentPage: page,
-        limitPerPage: limit,
-        nextLink: null,
-        prevLink: null
-      }
       if (result.length) {
         const data = await getCountAllSubCategoryModel(page)
         const { count } = data[0]
-        pageInfo.count = count
-        pageInfo.pages = Math.ceil(count / limit)
+        const pageInfo = pagination(req, count, '', 'sub-category')
 
-        const { pages, currentPage } = pageInfo
-
-        if (currentPage < pages) {
-          pageInfo.nextLink = `${APP_URL}/public/sub-category?${qs.stringify({ ...req.query, ...{ page: page + 1 } })}`
-        }
-
-        if (currentPage > 1) {
-          pageInfo.prevLink = `${APP_URL}/public/sub-category?${qs.stringify({ ...req.query, ...{ page: page - 1 } })}`
-        }
         return response(res, 'List of Sub Category', 200, true, { data: result, pageInfo })
       } else {
         return response(res, 'Sub Category not found', 400, false)
@@ -69,40 +64,46 @@ module.exports = {
 
   updateSubCategory: async (req, res) => {
     try {
-      const { id } = req.params
-      const { name } = req.body
+      const roleId = req.user.role_id
+      if (roleId === 2) {
+        const { id } = req.params
+        const { name } = req.body
 
-      const result = await updateSubCategoryModel([id, name])
-      if (result.affectedRows) {
-        return response(res, `Sub Category with id ${id} has been updated`)
+        const result = await updateSubCategoryModel([id, name])
+        if (result.affectedRows) {
+          return response(res, `Sub Category with id ${id} has been updated`, 200, true, req.body)
+        } else {
+          return response(res, 'Failed to update Sub Category', 400, false)
+        }
       } else {
-        return response(res, 'Failed to update Sub Category', 400, false)
+        return response(res, 'Forbidden access', 401, false)
       }
     } catch (err) {
-      return response(res, 'Internal server error', 500, false)
+      return response(res, 'Internal server error', 500, false, { error: err.message })
     }
   },
 
   deleteSubCategory: async (req, res) => {
     try {
-      const { id } = req.params
-      const data = await getDetailSubCategoryIDModel(id)
-      if (data.length) {
-        try {
+      const roleId = req.user.role_id
+      if (roleId === 2) {
+        const { id } = req.params
+        const data = await getDetailSubCategoryIDModel(id)
+        if (data.length) {
           const result = await deleteSubCategoryModel(id)
           if (result.affectedRows) {
             return response(res, `Sub Category with id ${id} has been deleted`)
           } else {
             return response(res, 'Sub Category cannot be deleted', 400, false)
           }
-        } catch (err) {
-          return response(res, 'Internal server error', 500, false)
+        } else {
+          return response(res, 'Sub Category not found', 404, false)
         }
       } else {
-        return response(res, 'Sub Category not found', 404, false)
+        return response(res, 'Forbidden access', 401, false)
       }
     } catch (err) {
-      return response(res, 'Internal server error', 500, false)
+      return response(res, 'Internal server error', 500, false, { error: err.message })
     }
   },
 
@@ -124,34 +125,15 @@ module.exports = {
       }
       const offset = (page - 1) * limit
       const result = await getDetailSubCategoryModel([id, limit, offset])
-      const pageInfo = {
-        count: 0,
-        pages: 0,
-        currentPage: page,
-        limitPerPage: limit,
-        nextLink: null,
-        prevLink: null,
-        numberOfData: result.length
-      }
       if (result.length) {
         const name = result[0].sub_category_name
         const data = await getCountSubCategoryModel([id, name])
-        const { count } = data[1]
-        pageInfo.count = count
-        pageInfo.pages = Math.ceil(count / limit)
+        const { count } = data[0]
+        const pageInfo = pagination(req, count, id, 'sub-category')
 
-        const { pages, currentPage } = pageInfo
-
-        if (currentPage < pages) {
-          pageInfo.nextLink = `${APP_URL}/public/sub-category/${id}?${qs.stringify({ ...req.query, ...{ page: page + 1 } })}`
-        }
-
-        if (currentPage > 1) {
-          pageInfo.prevLink = `${APP_URL}/public/sub-category/${id}?${qs.stringify({ ...req.query, ...{ page: page - 1 } })}`
-        }
         return response(res, `List of Sub Category ${name}`, 200, true, { data: result, pageInfo })
       } else {
-        return response(res, `There is no data on list of Sub Category with id ${id}`, 400, false, { pageInfo })
+        return response(res, `There is no data on list of Sub Category with id ${id}`, 400, false)
       }
     } catch (err) {
       return response(res, 'Internal server error', 500, false, { error: err.message })
